refactor(ui): clean up list view and extract reusable components
