--- EXSquaredTestApp/EXSquaredTestApp/App/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  EXSquaredTestApp
-//
-//  Created by bryangarcia on 5/2/25.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
